% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geneIDs.R
\name{check_IDs}
\alias{check_IDs}
\title{Check compatibility between gene IDs and organism}
\usage{
check_IDs(ids, organism)
}
\arguments{
\item{ids}{character vector of gene identifiers to be tested}

\item{organism}{organism, should be betwwen Arabidopsis thaliana,
Homo sapiens and Mus musculus.}
}
\value{
boolean, true if all of the gene IDs match the expected regex.
}
\description{
Check compatibility between gene IDs and organism
}
\examples{
data("abiotic_stresses")
check_IDs(rownames(abiotic_stresses$raw_counts),
organism = "Arabidopsis thaliana")
check_IDs(rownames(abiotic_stresses$raw_counts),
organism = "Homo sapiens")
}
