% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_analysis.R
\name{community_structure}
\alias{community_structure}
\title{Returns modules-communities memberships}
\usage{
community_structure(graph)
}
\arguments{
\item{graph}{igraph object, directed}
}
\value{
(named) vector
}
\description{
Returns modules-communities memberships
}
