% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{compute_pca}
\alias{compute_pca}
\title{compute_pca}
\usage{
compute_pca(data, kept_axes = 4)
}
\arguments{
\item{data}{normalized expression data with samples as columns and genes as rows.}

\item{kept_axes}{max number of component to keep.}
}
\description{
compute variables contributions to principal components,
as well as the PCA scree.
}
\examples{
data("abiotic_stresses")
pca <- compute_pca(abiotic_stresses$normalized_counts)
}
