% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{convert_from_agi}
\alias{convert_from_agi}
\title{For Arabdopsis thaliana, converts TAIR IDs to entrez IDs, 
gene symbol or description}
\usage{
convert_from_agi(ids, to = "entrez")
}
\arguments{
\item{ids}{vector of AGI gene IDs}

\item{to}{value in c("entre", "symbol", "name)}
}
\value{
named vector
}
\description{
For Arabdopsis thaliana, converts TAIR IDs to entrez IDs, 
gene symbol or description
}
\examples{
genes <- c("AT2G05940", "AT4G16480", "AT4G04570", "AT2G30130", "AT1G56300")
entrez_ids <- convert_from_agi(genes)
print(entrez_ids)
symbolss <- convert_from_agi(genes, to = "symbol")
print(symbols)
}
