% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{convert_from_ensembl}
\alias{convert_from_ensembl}
\title{For Homo sapiens, converts ensembl IDs to entrez IDs, 
symbol or name,}
\usage{
convert_from_ensembl(ids, to = "entrez")
}
\arguments{
\item{ids}{genes to convert, ensembl}

\item{to}{value in c("entrez", "symbol", "name")}
}
\value{
named list
}
\description{
For Homo sapiens, converts ensembl IDs to entrez IDs, 
symbol or name,
}
