% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{convert_from_ensembl_dm}
\alias{convert_from_ensembl_dm}
\title{For D. melanogaster, converts ensembl IDs to entrez IDs, 
symbol or name}
\usage{
convert_from_ensembl_dm(ids, to = "entrez")
}
\arguments{
\item{ids}{genes to convert, ensembl}

\item{to}{value in c("entrez", "symbol", "name")}
}
\value{
named vector
}
\description{
For D. melanogaster, converts ensembl IDs to entrez IDs, 
symbol or name
}
\examples{
if(require("org.Dm.eg.db")){
genes <- c("FBgn0000042", "FBgn0011224")
convert_from_ensembl_dm(genes)
convert_from_ensembl_dm(genes, to = "symbol")
convert_from_ensembl_dm(genes, to = "name")
}
}
