% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{convert_from_ensembl_mus}
\alias{convert_from_ensembl_mus}
\title{For Mus musculus, converts ensembl IDs to entrez IDs, 
symbol or name,}
\usage{
convert_from_ensembl_mus(ids, to = "entrez")
}
\arguments{
\item{ids}{genes to convert, ensembl}

\item{to}{value in c("entrez", "symbol", "name")}
}
\value{
named list
}
\description{
For Mus musculus, converts ensembl IDs to entrez IDs, 
symbol or name,
}
\examples{
if(require("org.Mm.eg.db")){
genes <- c("ENSMUSG00000000001", "ENSMUSG00000000049")
convert_from_ensembl_mus(genes)
convert_from_ensembl_mus(genes, to = "symbol")
convert_from_ensembl_mus(genes, to = "name")
}
}
