% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dea.R
\name{create_versus_design}
\alias{create_versus_design}
\title{create_versus_design}
\usage{
create_versus_design(experiment_design, reference_point, comparison_point)
}
\arguments{
\item{experiment_design}{a vector containing all the conditions of a studied dataset.}

\item{reference_point}{All the points in experiment_design to consider as reference}

\item{comparison_point}{All the points in experiment_design to consider as comparison points.}
}
\value{

}
\description{
create_versus_design
}
\examples{
create_versus_design(c("reference","condition1","condition2","condition3"), c("condition1", "condition2", "condition3"), "reference")
}
