% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data_At.R
\docType{data}
\name{demo_data_At}
\alias{demo_data_At}
\title{Expression matrix from Arabidopsis thaliana.}
\format{
A list with the following elements:
\describe{
 \item{raw_counts}{raw transcript aboundances as obtained after mapping and quantifying RNASeq reads (dataframe of integer positive values).}
 \item{conditions}{Experimental conditions of each sample, regardless of biological replicates. (character vector)}
 \item{design}{For each condition, the level of the corresponding factors.}
}
}
\source{
BPMP Institute, Supagro Montpellier
}
\usage{
demo_data_At
}
\description{
Expression matrix from Arabidopsis thaliana.
}
\examples{
{
 head(demo_data_At$raw_counts)
 demo_data_At$conditions
 demo_data_At$design
}
}
\keyword{datasets}
