% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_analysis.R
\name{describe_node}
\alias{describe_node}
\title{Targets and regulators of a gene in a network}
\usage{
describe_node(graph, node)
}
\arguments{
\item{graph}{igraph object}

\item{node}{gene to describe}
}
\value{
list with targets and regulators attributes
}
\description{
Targets and regulators of a gene in a network
}
\examples{
data("abiotic_stresses")
data("regulators_per_organism") 
mat <- abiotic_stresses$heat_DEGs_regulatory_links
network <- DIANE::network_thresholding(mat, n_edges = 600)
data <- network_data(network, regulators_per_organism[["Arabidopsis thaliana"]])
nodes <- data$nodes[order(-data$nodes$degree),]
gene_of_interest <- nodes$id[1] 
DIANE::describe_node(network, node =gene_of_interest)
}
