% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{draw_correlation_heatmap}
\alias{draw_correlation_heatmap}
\title{draw_correlation_heatmap}
\usage{
draw_correlation_heatmap(
  data = NULL,
  conds = unique(stringr::str_split_fixed(colnames(data), "_", 2)[, 1]),
  correlation_method = "pearson",
  low_color = "#db4760",
  high_color = "#4169e1",
  mid_color = "white",
  font_size = 3,
  limits = NULL
)
}
\arguments{
\item{data}{a matrix of count. This will be converted to log2+1}

\item{conds}{if NULL, shows all the conditions, else if character vector, shows only the required ones}

\item{correlation_method}{correlation method for the cor function. Can be pearson,
kendall or spearman.}

\item{low_color}{color for low correlation. Default : #db4760}

\item{high_color}{color for high correlation. Default : #4169e1}

\item{mid_color}{color for values in between high and low. Default : white}

\item{font_size}{font size for the text in cells.}

\item{limits}{vector with 3 elements, the lower limit, the middle, and the high.}
}
\description{
draw a heatmap of pearson correlation between conditions.
Display only two digits.
}
\examples{
data("abiotic_stresses")
draw_correlation_heatmap(abiotic_stresses$normalized_counts)
}
