% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_edge_testing.R
\name{draw_discarded_edges}
\alias{draw_discarded_edges}
\title{Draw network with removed edges in red}
\usage{
draw_discarded_edges(links, net_data)
}
\arguments{
\item{links}{dataframe of edges containing their adjusted pvalues, as returned 
by the \code{test_edges} function}

\item{net_data}{network data of the thresholded network, given by \code{DIANE::network_data()}}
}
\description{
Draw network with removed edges in red
}
\examples{
data(abiotic_stresses)
links <- abiotic_stresses$heat_edge_tests$links
net <- network_from_tests(links, fdr = 0.01)
net_data <- network_data(net, 
gene_info = gene_annotations$`Arabidopsis thaliana`, 
regulators = regulators_per_organism$`Arabidopsis thaliana`)
draw_discarded_edges(links, net_data)
}
