% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{draw_enrich_go}
\alias{draw_enrich_go}
\title{Plot the enriched go terms of an enrich_go result}
\usage{
draw_enrich_go(go_data, max_go = dim(go_data)[1])
}
\arguments{
\item{go_data}{dataframe returned by the enrich_go function,
containing for each significant GO, their gene count, gene symbols,
adjusted pvalue, description, etc.}

\item{max_go}{maximum number of GO terms to display. Default is the 
total significant GO number, but for legibility reasons, you may want 
to reduce that number}
}
\description{
Plot the enriched go terms of an enrich_go result
}
\examples{
\dontrun{
data("abiotic_stresses")
genes <- abiotic_stresses$heat_DEGs

genes <- get_locus(genes)
background <- get_locus(rownames(abiotic_stresses$normalized_counts))

genes <- convert_from_agi(genes)
background <- convert_from_agi(background)
go <- enrich_go(genes, background)
draw_enrich_go(go)
draw_enrich_go(go, max_go = 20)
}
}
