% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{draw_enrich_go_map}
\alias{draw_enrich_go_map}
\title{Draw GO enrichment map}
\usage{
draw_enrich_go_map(go)
}
\arguments{
\item{go}{dataframe as returned by \code{DIANE::enrich_go()} function.}
}
\description{
This plots each enriched GO term, with its pvalue and gene count 
as color and size aesthetics.
Two GO terms are linked by an edges if they share common genes
in the input gene list given for enrichment analysis. The width 
of the edge is proportionnal to that number of shared genes.
}
\examples{
data("abiotic_stresses")
data("gene_annotations")
genes <- convert_from_agi(get_locus(abiotic_stresses$heat_DEGs))
background <- convert_from_agi(get_locus(rownames(abiotic_stresses$normalized_counts)))
go <- enrich_go(genes, background)
draw_enrich_go_map(go)
}
