% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{draw_expression_levels}
\alias{draw_expression_levels}
\title{Draw gene expression levels}
\usage{
draw_expression_levels(
  data,
  genes,
  conds = unique(stringr::str_split_fixed(colnames(data), "_", 2)[, 1]),
  gene.name.size = 12,
  log2_count = FALSE,
  start_from_zero = FALSE,
  palette = getOption("DIANE.palette", "ggplot")
)
}
\arguments{
\item{data}{normalized expression dataframe, with genes as rownames and
conditions as colnames.}

\item{genes}{character vector of genes to be plotted (must be contained in
the rownames of data)}

\item{conds}{conditions to be shown on expression levels (must be contained in
the column names of data before the _rep suffix). Default : all conditions.}

\item{gene.name.size}{size of the facet plot title font for each gene. Default : 12}

\item{log2_count}{transform count using the log2 function. A pseudocount of 1 is added to
avoid negative values.}

\item{start_from_zero}{set the beginning of the y axis to 0.}

\item{palette}{categorical palette : "ggplot", "diane", "okabe", "npg",
"aaas", "lancet" or "nejm". Defaults to
getOption("DIANE.palette"), else "ggplot". Here it
colours the replicates.}
}
\description{
The normalized counts of the desired genes in the specified conditions are
shown. Please limit the number of input genes for readability reasons (up to 10 genes).
}
\examples{
genes <- sample(abiotic_stresses$heat_DEGs, 4)
conditions <- c("C", "M", 'H', 'SH')
DIANE::draw_expression_levels(abiotic_stresses$normalized_counts,
genes = genes, conds = conditions)
}
