% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glm.R
\name{draw_glm}
\alias{draw_glm}
\title{Plots the coefficients value of a Poisson generalized linear model}
\usage{
draw_glm(glm, palette = getOption("DIANE.palette", "ggplot"))
}
\arguments{
\item{glm}{glm object returned by \code{DIANE::fit_glm()}}

\item{palette}{categorical palette : "ggplot", "diane", "okabe", "npg",
"aaas", "lancet" or "nejm". Defaults to
getOption("DIANE.palette"), else "ggplot". Here it
colours the coefficients, not the conditions.}
}
\description{
Plots the coefficients value of a Poisson generalized linear model
}
\examples{
data("abiotic_stresses")
genes_cluster <- DIANE::get_genes_in_cluster(
abiotic_stresses$heat_DEGs_coseq_membership, cluster = 3)
glm <- DIANE::fit_glm(abiotic_stresses$normalized_counts, genes_cluster, 
abiotic_stresses$design)
draw_glm(glm)
}
