% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_MDS}
\alias{draw_MDS}
\title{Multi-dimensional scaling plot}
\usage{
draw_MDS(normalized.count, conditions = NULL)
}
\arguments{
\item{normalized.count}{data to plot for MDS}

\item{conditions}{if NULL (default, takes the split before the character '_' as condition names)
else, enter the conditions regardless of biological replicates, as a character vector. Its order should match the 
columns names of the expression matrix used to build the tcc object.}
}
\description{
Multi-dimensional scaling plot
}
\examples{
data("abiotic_stresses")
DIANE::draw_MDS(abiotic_stresses$normalized_counts)
}
