% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_analysis.R
\name{draw_network_degrees}
\alias{draw_network_degrees}
\title{Plots the histogram of in and out degrees, betweeness of 
regultors and target genes when relevant}
\usage{
draw_network_degrees(nodes, graph)
}
\arguments{
\item{nodes}{dataframe containing the nodes information}

\item{graph}{igraph object}
}
\description{
Plots the histogram of in and out degrees, betweeness of 
regultors and target genes when relevant
}
