% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_PCA}
\alias{draw_PCA}
\title{Draw PCA results}
\usage{
draw_PCA(data)
}
\arguments{
\item{data}{normalized expression data with samples as columns and genes as rows.}
}
\description{
Draws variables contributions to principal components,
as well as the PCA screeplot.
First to fourth principal components are shown.
}
\examples{
data("abiotic_stresses")
draw_PCA(abiotic_stresses$normalized_counts)
}
