% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{draw_PCA_legacy}
\alias{draw_PCA_legacy}
\title{Draw PCA results Legacy}
\usage{
draw_PCA_legacy(data)
}
\arguments{
\item{data}{normalized expression data with samples as columns and genes as rows.}
}
\description{
Draws variables contributions to principal components,
as well as the PCA screeplot.
First to fourth principal components are shown, except if there are
only 4 samples. In that case, 3 principal components are computed.
This function is the original DIANE function, which is preserved for 
reproductibility purpose.
}
\examples{
data("abiotic_stresses")
draw_PCA(abiotic_stresses$normalized_counts)
}
