% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{draw_pca_scree}
\alias{draw_pca_scree}
\title{draw_pca_scree}
\usage{
draw_pca_scree(pca)
}
\arguments{
\item{pca}{PCA data obtained from the compute_pca function.}
}
\description{
Display PCA scree plot, which shows contribution of all the
computed components.
}
\examples{
data("abiotic_stresses")
pca <- compute_pca(abiotic_stresses$normalized_counts)
draw_pca_scree(pca)
}
