% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{draw_profiles}
\alias{draw_profiles}
\title{Draw expression profiles of a clustering}
\usage{
draw_profiles(
  data,
  membership,
  conds,
  expression = "profiles",
  k = NULL,
  nrow = 3
)
}
\arguments{
\item{data}{normalized counts with genes as rownames and samples as columns}

\item{membership}{membership item of the coseq object returned by 
\code{run_coseq()} object}

\item{conds}{conditions on which to display clustering profiles.
Must be a unique vector containing the conditions you want to consider
for gene clustering, without the replicate information 
(string before the underscore in sample names)}

\item{expression}{if it is set to "profiles" (default), 
plots expression/sum(expression). if "counts", plots log(Counts+1)}

\item{k}{if NULL (default), plots all the clusters. Else, plot the clusters in the vetcor k.}

\item{nrow}{on how many rows display the cluster profiles if k is NULL}
}
\description{
Draw expression profiles of a clustering
}
\examples{
data("abiotic_stresses")
DIANE::draw_profiles(data = abiotic_stresses$normalized_counts, 
membership = abiotic_stresses$heat_DEGs_coseq_membership,
conds = unique(abiotic_stresses$conditions)) 
DIANE::draw_profiles(data = abiotic_stresses$normalized_counts, 
membership = abiotic_stresses$heat_DEGs_coseq_membership, 
conds = unique(abiotic_stresses$conditions), k = 3) 
}
