% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dea.R
\name{draw_raw_pvalue_histogram}
\alias{draw_raw_pvalue_histogram}
\title{Draw raw pvalue histogram}
\usage{
draw_raw_pvalue_histogram(tags, bins = 100, lfc = 0)
}
\arguments{
\item{tags}{returned by estimateDEGs, function, that is to say topTags from edgeR, 
used with \code{p.value = 1}}

\item{bins}{number of bar to display. Default is 100.}

\item{lfc}{absolute log fold change threshold for differentially expressed genes, default is 0.}
}
\value{
ggplot object corresponding to raw pvalue histogram.
}
\description{
Draw raw pvalue histogram
}
\examples{
data("abiotic_stresses")
tcc_object <- DIANE::normalize(abiotic_stresses$raw_counts, abiotic_stresses$conditions, 
iteration = FALSE)
tcc_object <- DIANE::filter_low_counts(tcc_object, 10*length(abiotic_stresses$conditions))
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = abiotic_stresses$conditions)
tags <- DIANE::estimateDEGs(fit, reference = "C", perturbation = "H", p.value = 1)
DIANE::draw_raw_pvalue_histogram(tags, bins=100, lfc = 0)
}
