% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dea.R
\name{draw_venn}
\alias{draw_venn}
\title{Venn diagram}
\usage{
draw_venn(gene_list)
}
\arguments{
\item{gene_list}{named list of genes}
}
\description{
Draws a Venn diagram between lists of DEGs.
Possible numbers of lists handled by the function are 2, 3 or 4.
}
\examples{
data("abiotic_stresses")
tcc_object <- DIANE::normalize(abiotic_stresses$raw_counts, abiotic_stresses$conditions, 
iteration = FALSE)
tcc_object <- DIANE::filter_low_counts(tcc_object, 10*length(abiotic_stresses$conditions))
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = abiotic_stresses$conditions)
genes_heat <- DIANE::estimateDEGs(fit, reference = "C", perturbation = "H",
 p.value = 0.05)$table$genes
genes_heat_mannitol <- DIANE::estimateDEGs(fit, reference = "C", perturbation = "MH", 
p.value = 0.05)$table$genes
draw_venn(list("C - H" = genes_heat, "C - MH" = genes_heat_mannitol))
}
