% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dea.R
\name{estimateDEGs}
\alias{estimateDEGs}
\title{estimateDEGs}
\usage{
estimateDEGs(fit, reference, perturbation)
}
\arguments{
\item{fit}{edgeR glmFit}

\item{reference}{condition being considered as "normal"}

\item{perturbation}{condition we can to compare to the reference}
}
\value{
tags dataframe
}
\description{
Given a glmFit, performs a log ratio test and
returns the genes that are differentially expressed
between the reference and perturbation conditions.
log ratios will be expressed as log(perturbation/reference)
}
