% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dea.R
\name{estimateDispersion}
\alias{estimateDispersion}
\title{estimateDispersion}
\usage{
estimateDispersion(tcc)
}
\arguments{
\item{tcc}{object countaining count, and norm factors}
}
\value{
fit object, containing teh fitted models
}
\description{
The function computes the common, trend, and tag-wise dispersions of the dataset.
It then fits a negative binomial model for each gene using default design : ~group+0 and GLMs.
}
