% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{filter_hts}
\alias{filter_hts}
\title{filter_hts}
\usage{
filter_hts(data, conditions)
}
\arguments{
\item{data}{data to be filtered to remove low count genes}

\item{conditions}{the sum of counts across all samples to be exceeded for a gene}
}
\value{
the filtered data
}
\description{
filter_hts
}
