% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{filter_low_count_condition_wise}
\alias{filter_low_count_condition_wise}
\title{Remove low count genes condition wise.}
\usage{
filter_low_count_condition_wise(tcc, threshold = 10, conditions = NULL)
}
\arguments{
\item{tcc}{data to be filtered to remove low count genes}

\item{threshold}{the sum of counts across all samples to be exceeded for a gene}

\item{conditions}{condition of each column of the data argument. Default is
all the conditions in the experiment. (as defined by the underscore prefixes).}
}
\value{
a TCC-Class object
}
\description{
Removes genes having a maximum median of counts accross all 
conditions lesser than the specified threshold. It returns un aupdated TCC 
object, which count element contains the filtered expression matrix. Compare
to the filter_low_counts function, it aims to preserve genes having medium
counts in a few batch of conditions. Code based on the filterLowCountGenes
function from the TCC package.
}
\examples{
data("abiotic_stresses")
tcc_object <- DIANE::normalize(abiotic_stresses$raw_counts, 
abiotic_stresses$conditions, iteration = FALSE)
threshold = 10*length(abiotic_stresses$conditions)
tcc_object <- DIANE::filter_low_count_condition_wise(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
}
