% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{filter_low_counts}
\alias{filter_low_counts}
\title{filter_low_counts}
\usage{
filter_low_counts(tcc, thr)
}
\arguments{
\item{tcc}{data to be filtered to remove low count genes}

\item{thr}{the sum of counts across all samples to be exceeded for a gene}
}
\value{
a TCC-Class object containing the filtered data
}
\description{
filter_low_counts
}
\examples{
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)
threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
}
