% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{filter_sum}
\alias{filter_sum}
\title{filter_sum}
\usage{
filter_sum(tcc_, thr)
}
\arguments{
\item{tcc_}{data to be filtered to remove low count genes}

\item{thr}{the sum of counts across all samples to be exceeded for a gene}
}
\value{
the filtered data
}
\description{
filter_sum
}
