% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_GENIE3_oob.R
\name{GENIE3OOB}
\alias{GENIE3OOB}
\title{GENIE3OOB}
\usage{
GENIE3OOB(
  exprMatrix,
  regulators = NULL,
  targets = NULL,
  K = "sqrt",
  nTrees = 1000,
  nCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{exprMatrix}{Expression matrix (genes x samples). Every row is a gene, every column is a sample.}

\item{regulators}{Subset of genes used as candidate regulators. Must be either a
vector of indices, e.g. \code{c(1,5,6,7)}, or a vector of gene names, e.g.
\code{c("at_12377", "at_10912")}. The default value NULL means that all the genes are used as candidate regulators.}

\item{targets}{Subset of genes to which potential regulators will be calculated. Must be either
a vector of indices, e.g. \code{c(1,5,6,7)}, or a vector of gene names, e.g. \code{c("at_12377", "at_10912")}.
If NULL (default), regulators will be calculated for all genes in the input matrix.}

\item{nTrees}{Number of trees in an ensemble for each target gene. Default: 1000.}

\item{nCores}{Number of cores to use for parallel computing. Default: 1.}

\item{verbose}{If set to TRUE, a feedback on the progress of the calculations is given. Default: TRUE}
}
\value{
matrix containing oob importances for all TF-gene pairs
}
\description{
\code{GENIE3OOB} Infers a gene regulatory network (in the form of a pvalues-filled adjacency matrix)
from expression data, using ensembles of regression trees.
}
