% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_regressors_grouping.R
\name{get_correlation}
\alias{get_correlation}
\title{Get correlation}
\usage{
get_correlation(pair, normalized.count)
}
\arguments{
\item{pair}{character of two gene IDs sepearte by a white space}

\item{normalized.count}{normalized expression matrix containing
the IDs of the pair}
}
\value{
spearman correlation bewteen the two expression vectors
}
\description{
Get correlation
}
