% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glm.R
\name{get_factors_from_conditions}
\alias{get_factors_from_conditions}
\title{Returns the factors that 
don't stay at the same level between 
between the specified conditions}
\usage{
get_factors_from_conditions(conditions, design)
}
\arguments{
\item{conditions}{character vetcor of conditions (must be in design rownames)}

\item{design}{design dataframe}
}
\value{
character vector of "active", or "perturbated" factors
}
\description{
Returns the factors that 
don't stay at the same level between 
between the specified conditions
}
