% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geneIDs.R
\name{get_locus}
\alias{get_locus}
\title{Get gene IDs loci, without splicing information}
\usage{
get_locus(gene_ids, unique = TRUE)
}
\arguments{
\item{gene_ids}{list of gene ids with splice variants information}

\item{unique}{boolean, weather or not to return unique locus vector}
}
\value{
character vector of gene IDs without alternative transcripts 
information
}
\description{
Get the locus IDs from splice variants IDs
(remove the .1, .2 from a list of gene IDs)
}
\examples{
splice_variants <- rownames(abiotic_stresses$normalized_counts)[1:20]
get_locus(splice_variants)
}
