% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_edge_testing.R
\name{get_nEdges}
\alias{get_nEdges}
\title{Get the number of a edges}
\usage{
get_nEdges(density, nGenes, nRegulators)
}
\arguments{
\item{density}{network density (usually between 0.001 and 0.1)}

\item{nGenes}{total number of genes}

\item{nRegulators}{number of genes that are regulators}
}
\value{
number of edges
}
\description{
Returns the number of edges of a Gene Regulatory
Network from its density.
}
