% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{interGO}
\alias{interGO}
\title{Number of shared genes between Two GO terms}
\usage{
interGO(go_pair, go_table)
}
\arguments{
\item{go_pair}{}

\item{go_table}{go as returned by enrich_go}
}
\value{
integer
}
\description{
Number of shared genes between Two GO terms
}
