% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_inference.R
\name{network_inference}
\alias{network_inference}
\title{Genie3 regulatory importances estimation}
\usage{
network_inference(
  normalized.count,
  conds,
  regressors,
  targets,
  nTrees = 5000,
  nCores = 1
)
}
\arguments{
\item{normalized.count}{normalized expression matrix containing the regressors and target genes}

\item{conds}{condition names to be used in the inference}

\item{regressors}{genes to be taken as regressors during the inference procedures (regulator genes)}

\item{targets}{genes to be included in the inferred network. Regressors can also be in the targets}

\item{nTrees}{Number of trees by Random Forest}

\item{nCores}{Number of CPU cores to use during the procedure}
}
\value{
matrix object
}
\description{
Genie3 regulatory importances estimation
}
