% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{normalize}
\alias{normalize}
\title{normalize raw count data}
\usage{
normalize(
  data,
  conditions,
  norm_method = "tmm",
  deg_method = "edgeR",
  fdr = 0.01
)
}
\arguments{
\item{data}{raw counts to be normalized (data frame or matrix)}

\item{conditions}{condition of each column of the data argument}

\item{norm_method}{method used for normalisation, between tmm or deseq2}

\item{deg_method}{method used for deg detection, between edgeR ou deseq2}

\item{fdr}{pvalue threshold for adjusted pvalues for degs}
}
\value{
the normalized data
}
\description{
normalize raw count data
}
