% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organisms.R
\docType{data}
\name{organisms}
\alias{organisms}
\title{Custom organisms}
\format{
A list. Each item of the list correspond to an organism. for each of
them, some fields can be set. 
\describe{
 \item{annotation}{Gene annotation }
 \item{regulators}{Regulators. Can be left empty}
 \item{gene_exemple}{A gene exemple that will be displayed inside of the online interface }
 \item{go}{Gene to GO mapping}
 \item{informations}{Free fields, with informations about how data have been collected.
 Can contain any informations you think are worth of being display to end users.
 Can contain either a string, or a list, where each elements of the list 
 correspond to a specific topic.}
 }
}
\usage{
organisms
}
\description{
This dataset contains custom organism data. It has been made to be easy to edit
and allow to add any kind of custom organism to DIANE as long as you follow the
specified nomenclature. This is especially usefull for the server version of
DIANE, who parses this dataset.
}
\examples{
{
 print(names(organisms))
 print(head(organisms[["Lupinus albus"]][["annotation"]]))
}
}
\keyword{datasets}
