% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulators_per_organism.R
\docType{data}
\name{regulators_per_organism}
\alias{regulators_per_organism}
\title{List of regulator genes for a number of model organisms}
\format{
A named list with the following elements:
\describe{
 \item{Arabidopsis thaliana}{character vector of the known regulators in Arabidopsis thaliana}
 \item{Homo sapiens}{character vector of the known transcription factors in Human}
}
}
\source{
\describe{
 \item{For Arabidopsis thaliana} {http://plntfdb.bio.uni-potsdam.de/v3.0/index.php?sp_id=ATH}
 \item{For Human}{http://humantfs.ccbr.utoronto.ca/download.php}
 }
}
\usage{
regulators_per_organism
}
\description{
List of regulator genes for a number of model organisms
}
\examples{
{
 print(head(regulators_per_organism[["Arabidopsis thaliana"]]))
 print(head(regulators_per_organism[["Homo sapiens"]]))
}
}
\keyword{datasets}
