% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{run_coseq}
\alias{run_coseq}
\title{run_coseq}
\usage{
run_coseq(conds, genes, data, K = 6:12)
}
\arguments{
\item{conds}{Condition names to be used for clustering}

\item{genes}{Genes used as an input for the clustering}

\item{data}{normalized counts (MUST be normalized!)}

\item{K}{number of clusters range}
}
\value{
named list containing the coseq run result as "model", and the cluster membership 
for each gene as "membership"
}
\description{
run_coseq
}
\examples{
data("abiotic_stresses")
genes <- abiotic_stresses$heat_DEGs
clustering <- DIANE::run_coseq(conds = unique(abiotic_stresses$conditions), 
data = abiotic_stresses$normalized_counts, genes = genes, K = 6:9)
}
