% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geneIDs.R
\name{aggregate_splice_variants}
\alias{aggregate_splice_variants}
\title{Aggregate splice variants}
\usage{
aggregate_splice_variants(data)
}
\arguments{
\item{data}{expression dataframe with splice variants as rownames
and samples as columns}
}
\value{
dataframe with aggregated rows
}
\description{
Merge all splice variants of an expression dataset into 
unique locus, unaware of alternative splicing, by summing
all variants of the same gene
}
\examples{
data("abiotic_stresses")
head(aggregate_splice_variants(abiotic_stresses$normalized_counts))
}
