% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{aggregate_splice_variants}
\alias{aggregate_splice_variants}
\title{Merge all splice variants of an expression dataset into 
unique locus, unaware of alternative splicing, by summing
all variants for the same gene}
\usage{
aggregate_splice_variants(data)
}
\arguments{
\item{data}{dataframe : expression data with splice variants as rownames}
}
\value{
dataframe with agregated rows
}
\description{
Merge all splice variants of an expression dataset into 
unique locus, unaware of alternative splicing, by summing
all variants for the same gene
}
\examples{
data("abiotic_stresses")
d <- aggregate_splice_variants(abiotic_stresses$normalized_counts)
}
