% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{convert_from_ensembl}
\alias{convert_from_ensembl}
\title{For H. sapiens, converts ensembl IDs to entrez IDs, 
symbol or name}
\usage{
convert_from_ensembl(ids, to = "entrez")
}
\arguments{
\item{ids}{genes to convert, ensembl}

\item{to}{value in c("entrez", "symbol", "name")}
}
\value{
named vector
}
\description{
For H. sapiens, converts ensembl IDs to entrez IDs, 
symbol or name
}
\examples{
if(require("org.Hs.eg.db")){
genes <- c("ENSG00000000003", "ENSG00000003989", "ENSG00000005884", "ENSG00000007168")
convert_from_ensembl(genes)
convert_from_ensembl(genes, to = "symbol")
}

}
