% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_analysis.R
\name{describe_node}
\alias{describe_node}
\title{Targets and regulators of a gene in a network}
\usage{
describe_node(graph, node)
}
\arguments{
\item{graph}{igraph object}

\item{node}{gene to describe}
}
\value{
list with targets and regulators attributes
}
\description{
Targets and regulators of a gene in a network
}
