% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{draw_coseq_run}
\alias{draw_coseq_run}
\title{Displays clustering results and quality}
\usage{
draw_coseq_run(clustering_run, plot = "ICL")
}
\arguments{
\item{clustering_run}{result of a coseq run}

\item{plot}{plot to display, either Integrated Complete Likelihood, or barplots
of the posterior probabilities for the clustering. Value must be "ICL" or "barplots".}
}
\value{
plot describing the quality of the clustering process
}
\description{
Displays clustering results and quality
}
\examples{
data("abiotic_stresses")
genes <- abiotic_stresses$heat_DEGs
clustering <- DIANE::run_coseq(conds = c("C", "H", "SH", "MH", "SMH"), 
data = abiotic_stresses$normalized_counts, genes = genes, K = 6:9)
DIANE::draw_coseq_run(clustering$model, plot = "barplots")
DIANE::draw_coseq_run(clustering$model, plot = "ICL")
}
