% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{draw_coseq_run}
\alias{draw_coseq_run}
\title{draw_coseq_run}
\usage{
draw_coseq_run(run_pois, plot = "ICL")
}
\arguments{
\item{run_pois}{result of a coseq run}
}
\value{
plot describing the quality of the clustering process
}
\description{
draw_coseq_run
}
