% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{draw_coseq_run}
\alias{draw_coseq_run}
\title{draw_coseq_run : displays the indications of clustering}
\usage{
draw_coseq_run(run_pois, plot = "ICL")
}
\arguments{
\item{run_pois}{result of a coseq run}

\item{plot}{plot to display, eather integrated Complete Likelihood, or barplots
of the posterior probabilities for the clustering. Value must be "ICL" or "barplots".}
}
\value{
plot describing the quality of the clustering process
}
\description{
draw_coseq_run : displays the indications of clustering
}
\examples{
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)
threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)
topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes
clustering <- DIANE::run_coseq(conds = unique(demo_data_At$conditions), data = normalized_counts, genes = genes, K = 6:9)
DIANE::draw_coseq_run(clustering$model, plot = "barplots")
DIANE::draw_coseq_run(clustering$model, plot = "ICL")
}
