% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{draw_distributions}
\alias{draw_distributions}
\title{Draw distributions of expression data}
\usage{
draw_distributions(data, type = "boxplot")
}
\arguments{
\item{data}{expression dataframe, with samples as columns and genes as rows}

\item{type}{one of the following : boxplot, density, density_ridges}
}
\description{
Draw distributions of expression data
}
\examples{
data("abiotic_stresses")
DIANE::draw_distributions(abiotic_stresses$normalized_counts, type = "boxplot")
DIANE::draw_distributions(abiotic_stresses$raw_counts)
}
