% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_distributions}
\alias{draw_distributions}
\title{Draw distributions of expression data}
\usage{
draw_distributions(data, boxplot = TRUE)
}
\arguments{
\item{data}{expression dataframe, with samples as columns and genes as rows}

\item{boxplot}{if TRUE, plot each sample as a boxplot, else, it is shown as distributions}
}
\description{
Draw distributions of expression data
}
\examples{
data("abiotic_stresses")
DIANE::draw_distributions(abiotic_stresses$normalized_counts, boxplot = FALSE)
DIANE::draw_distributions(abiotic_stresses$raw_counts)
}
