% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_distributions}
\alias{draw_distributions}
\title{draw_distributions of count data}
\usage{
draw_distributions(data, boxplot = TRUE)
}
\arguments{
\item{data}{count data, as a numeric dataframe with condition names as columns and genes as rows}

\item{boxplot}{if TRUE, plot each sample as a boxplot, else, it is shown as a violin plot}
}
\description{
draw_distributions of count data
}
\examples{
data("demo_data_At")
DIANE::draw_distributions(demo_data_At$raw_counts, boxplot = FALSE)
}
