% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_distributions}
\alias{draw_distributions}
\title{draw_distributions of count data}
\usage{
draw_distributions(data, boxplot = T)
}
\arguments{
\item{data}{count data}

\item{boxplot}{if TRUE, plot each sample as a boxplot, else, as a violin plot}
}
\value{
plot
}
\description{
draw_distributions of count data
}
