% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{draw_enrich_go}
\alias{draw_enrich_go}
\title{Plot the enriched go terms of an enrich_go result}
\usage{
draw_enrich_go(go_data, max_go = dim(go_data)[1])
}
\arguments{
\item{go_data}{dataframe returned by the enrich_go function,
containing for each significant GO, their gene count, gene symbols,
adjusted pvalue, description, etc.}

\item{max_go}{maximum number of GO terms to display. Default is the 
total significant GO number, but for legibility reasons, you may want 
to reduce that number}
}
\description{
Plot the enriched go terms of an enrich_go result
}
\examples{
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, 
iteration = FALSE)
threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object,
 conditions = demo_data_At$conditions)
topTags <- DIANE::estimateDEGs(fit, reference = "cNF", 
perturbation = "cnF", p.value = 0.01)
# interest and background gene sets as entrez ids
genes <- convert_from_agi(topTags$table$genes)
background <- convert_from_agi(rownames(normalized_counts))

go <- enrich_go(genes, background)
draw_enrich-go(go)
draw_enrich-go(go, max_go = 20)
}
