% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glm.R
\name{draw_glm}
\alias{draw_glm}
\title{Plots the value of a Poisson generalized linear model}
\usage{
draw_glm(glm)
}
\arguments{
\item{glm}{glm object returned by ```DIANE::fit_glm()```}
}
\description{
Plots the value of a Poisson generalized linear model
}
