% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glm.R
\name{draw_glm}
\alias{draw_glm}
\title{Plots the value of a Poisson generalized linear model coefficients}
\usage{
draw_glm(glm)
}
\arguments{
\item{glm}{glm object returned by ```DIANE::fit_glm()```}
}
\description{
Plots the value of a Poisson generalized linear model coefficients
}
\examples{
data("abiotic_stresses")
genes_cluster <- DIANE::get_genes_in_cluster(
abiotic_stresses$heat_DEGs_coseq_membership, cluster = 3)
glm <- DIANE::fit_glm(abiotic_stresses$normalized_counts, genes_cluster, 
abiotic_stresses$design)
draw_glm(glm)
}
