% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glm.R
\name{draw_glm}
\alias{draw_glm}
\title{Plots the value of a Poisson generalized linear model}
\usage{
draw_glm(glm)
}
\arguments{
\item{glm}{glm object returned by ```DIANE::fit_glm()```}
}
\description{
Plots the value of a Poisson generalized linear model
}
\examples{
library(DIANE)
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)
threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)
topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes
clustering <- DIANE::run_coseq(conds = unique(demo_data_At$conditions), 
data = normalized_counts, genes = genes, K = 6:9)
genes_cluster <- DIANE::get_genes_in_cluster(clustering$membership, cluster = 3)
glm <- DIANE::fit_glm(normalized_counts, genes_cluster, demo_data_At$design)
draw_glm(glm)
}
