% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw heatmap}
\usage{
draw_heatmap(
  normalized.count,
  subset = "random",
  show_rownames = F,
  title = "Random preview heatmap of log(expression)"
)
}
\arguments{
\item{normalized.count}{}

\item{subset}{}

\item{show_rownames}{}

\item{title}{}
}
\value{
plot of the heatmap
}
\description{
Draw heatmap
}
