% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw expression heatmap}
\usage{
draw_heatmap(
  data,
  subset = NULL,
  show_rownames = FALSE,
  title = "Expression dataset",
  log = TRUE,
  profiles = FALSE,
  conditions = NULL
)
}
\arguments{
\item{data}{expression dataframe, with genes as rownames and samples as columns}

\item{subset}{subset of genes to be display}

\item{show_rownames}{show rownames or not}

\item{title}{plot title}

\item{log}{Show log(expression+1) in the heatmap if TRUE, expression if FALSE}

\item{profiles}{Show expression/mean(expression) for each gene if TRUE, expression if FALSE}

\item{conditions}{if NULL, shows all the conditions, else if character vector, shows only the required ones}
}
\description{
Draw expression heatmap
}
\examples{
data("abiotic_stresses")
DIANE::draw_heatmap(abiotic_stresses$normalized_counts, subset = abiotic_stresses$heat_DEGs,
title = "Log expression for DE genes under heat stress")
}
