% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw heatmap}
\usage{
draw_heatmap(
  data,
  subset = NULL,
  show_rownames = F,
  title = "Random preview heatmap",
  log = TRUE,
  profiles = FALSE,
  conditions = NULL
)
}
\arguments{
\item{data}{data to plot}

\item{subset}{subset of rows to display}

\item{show_rownames}{show rownames or not}

\item{title}{plot title}

\item{log}{Show log(expression+1) in the heatmap if TRUE, expression if FALSE}

\item{profiles}{Show expression/mean(expression) for each gene if TRUE, expression if FALSE}

\item{conditions}{if NULL, shows all the conditions, else if character vector, shows only the required ones}
}
\value{
plot of the heatmap
}
\description{
Draw heatmap
}
