% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_MDS}
\alias{draw_MDS}
\title{Multi-dimensional scaling plot}
\usage{
draw_MDS(normalized.count)
}
\arguments{
\item{normalized.count}{data to plot for MDS}
}
\description{
Multi-dimensional scaling plot
}
\examples{
data("abiotic_stresses")
DIANE::draw_MDS(abiotic_stresses$normalized_counts)
}
