% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_inference.R
\name{draw_network}
\alias{draw_network}
\title{Displays an interavtive network view}
\usage{
draw_network(nodes, edges)
}
\arguments{
\item{nodes}{dataframe containing nodes}

\item{edges}{dataframe containing edges}
}
\description{
Displays an interavtive network view
}
