% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_inference.R
\name{draw_network}
\alias{draw_network}
\title{Interactive network view}
\usage{
draw_network(nodes, edges)
}
\arguments{
\item{nodes}{dataframe containing nodes, from \code{DIANE::network_data()$nodes}}

\item{edges}{dataframe containing edges, from \code{DIANE::network_data()$edges}}
}
\description{
Visualize the GRN with regulators as green square nodes,
and targets as gray discs.
}
\examples{
\dontrun{
data("abiotic_stresses")
data("regulators_per_organism")
genes <- get_locus(abiotic_stresses$heat_DEGs)
# mat was inferred using the function network_inference
mat <- abiotic_stresses$heat_DEGs_regulatory_links
network <- DIANE::network_thresholding(mat, n_edges = length(genes))
data <- network_data(network, regulators_per_organism[["Arabidopsis thaliana"]])
DIANE::draw_network(data$nodes, data$edges)}
}
