% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_analysis.R
\name{draw_network_degrees}
\alias{draw_network_degrees}
\title{Plots the histogram of in and out degrees, betweeness of 
regultors and target genes.}
\usage{
draw_network_degrees(nodes, graph)
}
\arguments{
\item{nodes}{dataframe containing the nodes information}

\item{graph}{igraph object}
}
\description{
Plots the histogram of in and out degrees, betweeness of 
regultors and target genes.
}
\examples{
data("demo_data_At")
data("regulators_per_organism")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)
threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)
topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes

regressors <- intersect(genes, regulators_per_organism[["Arabidopsis thaliana"]])
mat <- DIANE::network_inference(normalized_counts, conds = demo_data_At$conditions, 
targets = genes, regressors = regressors)
network <- DIANE::network_thresholding(mat, n_edges = length(genes))

data <- network_data(network, regulators_per_organism[["Arabidopsis thaliana"]])
DIANE::draw_network_degrees(data$nodes, network)
}
