% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_analysis.R
\name{draw_network_degrees}
\alias{draw_network_degrees}
\title{Plots the histogram of in and out degrees, betweeness of 
regultors and target genes.}
\usage{
draw_network_degrees(nodes, graph)
}
\arguments{
\item{nodes}{dataframe containing the nodes information}

\item{graph}{igraph object}
}
\description{
Plots the histogram of in and out degrees, betweeness of 
regultors and target genes.
}
\examples{
data("abiotic_stresses")
data("regulators_per_organism")

mat <- abiotic_stresses$heat_DEGs_regulatory_links
network <- DIANE::network_thresholding(mat, n_edges = length(abiotic_stresses$heat_DEGs))

data <- network_data(network, regulators_per_organism[["Arabidopsis thaliana"]])
DIANE::draw_network_degrees(data$nodes, network)
}
