% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_heatmap.R
\name{draw_PCA}
\alias{draw_PCA}
\title{Draw variables along principal PCA components}
\usage{
draw_PCA(data)
}
\arguments{
\item{data}{normalized expression data}
}
\description{
Draw variables along principal PCA components
}
\examples{
data("demo_data_At")
tcc_object <-
DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)
threshold = 10 * length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
draw_PCA(normalized_counts)
}
