% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{draw_profiles}
\alias{draw_profiles}
\title{Draw profiles of a clustering}
\usage{
draw_profiles(
  data,
  membership,
  conds,
  expression = "profiles",
  k = NULL,
  nrow = 3
)
}
\arguments{
\item{data}{normalized counts}

\item{membership}{membership item of the coseq object returned by run_coseq object}

\item{conds}{conditions on which to perform clustering, ingoring the others}

\item{expression}{if it is set to "profiles" (default), plots expression/sum(expression). if "counts", plots log(Counts+1)}

\item{k}{if NULL (default), plot all the clusters. Else, plot the clusters in the vetcor k.}

\item{nrow}{on how many rows display the cluster profiles if k is NULL}
}
\value{
ggplot
}
\description{
Draw profiles of a clustering
}
