% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{draw_specific_pca}
\alias{draw_specific_pca}
\title{draw_specific_pca}
\usage{
draw_specific_pca(pca, component_1, component_2, legend = TRUE)
}
\arguments{
\item{pca}{PCA data obtained from compute_pca function.}

\item{component_1}{First component to plot}

\item{component_2}{Second component to plot}

\item{legend}{Display legend on the plot.}
}
\description{
plot PCA results.
}
\examples{
data("abiotic_stresses")
pca <- compute_pca(abiotic_stresses$normalized_counts)
draw_specific_pca(pca, 1, 2)
}
