% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{enrich_go_custom}
\alias{enrich_go_custom}
\title{Enriched ontologies for a custom organism}
\usage{
enrich_go_custom(
  genes,
  universe = genes_to_GO[, 1],
  genes_to_GO,
  qvalue = 0.1,
  pvalue = 0.05,
  GO_type = "BP"
)
}
\arguments{
\item{genes}{list of gene IDs, as present in the first column od genes_to_GO}

\item{universe}{list of gene IDs as background, default is set to all the genes 
contained in genesto_GO}

\item{genes_to_GO}{dataframe with gene IDs as first column, and GO IDs as second column.
gene IDs must be the IDs present in your gene expresion file}

\item{qvalue}{qvalue cutoff for enriched GO terms, default to 0.1}

\item{pvalue}{qvalue cutoff for enriched GO terms, default to 0.05}

\item{GO_type}{character between BP (Biological Process), CC(Cellular Component) 
or MF (Molecular Function), depending on the GO type
to use in the analysis}
}
\value{
dataframe containing enriched go terms and their description
}
\description{
Enriched custom Gene Ontology terms in a set of genes, from a dataframe
giving Go terms to genes matching
}
