% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dea.R
\name{estimateDEGs}
\alias{estimateDEGs}
\title{estimateDEGs}
\usage{
estimateDEGs(fit, reference, perturbation, p.value = 1, lfc = 0)
}
\arguments{
\item{fit}{edgeR glmFit}

\item{reference}{condition being considered as teh reference for differential analysis}

\item{perturbation}{condition we compared to the reference for differential analysis}

\item{p.value}{numeric cutoff value for adjusted p-values. Only tags with adjusted p-values equal or 
lower than specified are returned}

\item{lfc}{minimal absolute log fold change required for a gene to be considered as 
differnetially expressed.}
}
\value{
topTags object, which table element contains DEGs dataframe.
}
\description{
Given a glmFit, performs a log ratio test using edgeR function glmLRT and
returns the genes that are differentially expressed
between the reference and perturbation conditions.
log ratios are expressed as log(perturbation/reference)
}
\examples{
data("abiotic_stresses")
tcc_object <- DIANE::normalize(abiotic_stresses$raw_counts, 
abiotic_stresses$conditions, iteration = FALSE)
threshold = 10*length(abiotic_stresses$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
fit <- DIANE::estimateDispersion(tcc = tcc_object, 
conditions = abiotic_stresses$conditions)
topTags <- DIANE::estimateDEGs(fit, reference = "C", perturbation = "H", p.value = 0.01)
DEGs <- topTags$table
}
