% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dea.R
\name{estimateDispersion}
\alias{estimateDispersion}
\title{estimateDispersion}
\usage{
estimateDispersion(tcc, conditions = NULL)
}
\arguments{
\item{tcc}{TCC-class object countaining counts, and norm factors, such as obtained after 
DIANE::normalize() and DIANE::filter_low_counts()}

\item{conditions}{if NULL (default), takes the split before the character '_' as condition names.
else, the conditions specified by the user, as a character vector. Its order should match the 
columns names of the expression matrix used to build the tcc object.}
}
\value{
glmFit
}
\description{
This function computes the common, trended, and tag-wise dispersions of 
the gene expression dataset. Once, the dispersion is estimated,
it fits a negative binomial model for each gene using default design : ~group+0.
In this configuration, the log average expression of each gene is approximated
by a linear combination of each of the conditions.
It returns a the value of a glmFit object from the package edgeR, containing the attributes
of the tcc object given, plus fitted values and GLM coefficients, among other indicators of the 
fitting procedure.
}
\section{# Note}{

The return value is meant to be used directly as the parameter of
DIANE::estimateDEGs()
}

\examples{
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)
threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)

}
