% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{filter_low_counts}
\alias{filter_low_counts}
\title{Remove low count genes}
\usage{
filter_low_counts(tcc, thr)
}
\arguments{
\item{tcc}{data to be filtered to remove low count genes}

\item{thr}{the sum of counts across all samples to be exceeded for a gene}
}
\value{
a TCC-Class object
}
\description{
Removes genes having a sum of counts accross all sample
inferior to the specified threshold. It returns un aupdated TCC object,
which count element contains the filtered expression matrix.
}
\examples{
data("abiotic_stresses")
tcc_object <- DIANE::normalize(abiotic_stresses$raw_counts, 
abiotic_stresses$conditions, iteration = FALSE)
threshold = 10*length(abiotic_stresses$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
}
