% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_glm.R
\name{fit_glm}
\alias{fit_glm}
\title{Fits a Poisson generalized linear model to a set of genes}
\usage{
fit_glm(normalized_counts, genes, design, factors = colnames(design))
}
\arguments{
\item{normalized_counts}{normalized counts}

\item{genes}{genes belonging to a specific expression-based clusters}

\item{design}{experimental design as a dataframe}

\item{factors}{factors to use for the fit (defalut is
all the factors of the design)}
}
\value{
glm object
}
\description{
more to come
}
