% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\docType{data}
\name{gene_annotations}
\alias{gene_annotations}
\title{Gene annotations for supported organisms in DIANE}
\format{
A named list with the following elements:
\describe{
 \item{Arabidopsis thaliana}{Dataframe with AGI terms as rownames, and columns two columns :
 description and name.}
}
}
\source{
\describe{
 \item{For Arabidopsis thaliana}{\url{http://bioconductor.org/packages/release/data/annotation/html/org.At.tair.db.html}
 {TAIR10 via the bioconductor package org.At.tair.db}}
 }
}
\usage{
gene_annotations
}
\description{
For each gene identifier, are provided its common name and description.
}
\examples{
{
 print(head(gene_annotations[["Arabidopsis thaliana"]]))
}
}
\keyword{datasets}
