% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_go.R
\name{get_gene_information}
\alias{get_gene_information}
\title{Gives gene information (common name and description) for a specific organism}
\usage{
get_gene_information(ids, organism)
}
\arguments{
\item{organism}{value in c("Arabidopsis thaliana", "Homo sapiens")}

\item{genes}{vector of genes, AGI for Arabidopsis and ensembl for Human}
}
\value{
a dataframe with input genes as rownames, and columns label and desciption
}
\description{
Gives gene information (common name and description) for a specific organism
}
\examples{
genes <-  c("AT2G05940", "AT4G16480", "AT4G04570", "AT2G30130", "AT1G56300")
get_gene_information(genes, organism = "Arabidopsis thaliana")

}
