% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{get_genes_in_cluster}
\alias{get_genes_in_cluster}
\title{get_genes_in_cluster}
\usage{
get_genes_in_cluster(membership, cluster)
}
\arguments{
\item{membership}{named vector from run_coseq}

\item{cluster}{desired cluster number}
}
\value{
genes belonging to the desired cluster
}
\description{
get_genes_in_cluster
}
\examples{
data("abiotic_stresses")
get_genes_in_cluster(abiotic_stresses$heat_DEGs_coseq_membership, 3)
}
