% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{get_locus}
\alias{get_locus}
\title{Get the locus ids from splice variants ids}
\usage{
get_locus(gene_ids)
}
\arguments{
\item{gene_ids}{list of gene ids with splice variants information}
}
\value{
character vector
}
\description{
Get the locus ids from splice variants ids
}
