% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{get_locus}
\alias{get_locus}
\title{Get the locus ids from splice variants ids}
\usage{
get_locus(gene_ids, unique = TRUE)
}
\arguments{
\item{gene_ids}{list of gene ids with splice variants information}

\item{unique}{boolean, weather or not to return unique locus vector}
}
\value{
character vector
}
\description{
Get the locus ids from splice variants ids
}
\examples{
splice_variants <- rownames(abiotic_stresses$normalized_counts)[1:20]
get_locus(splice_variants)
}
