% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_inference.R
\name{network_data}
\alias{network_data}
\title{Get data compatible with visNetwork from igraph object}
\usage{
network_data(graph, regulators)
}
\arguments{
\item{graph}{igraph object}

\item{regulators}{list of regulators}
}
\value{
list of dataframes containing nodes and edges information
}
\description{
Get data compatible with visNetwork from igraph object
}
