% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_inference.R
\name{network_data}
\alias{network_data}
\title{Creates data compatible with visNetwork from an igraph object}
\usage{
network_data(graph, regulators, gene_info = NULL)
}
\arguments{
\item{graph}{igraph object}

\item{regulators}{list of regulators, so they can be marked
as special nodes in the network}

\item{gene_info}{dataframe with gene IDs as rownames, containing
additional info on genes, in columns named label and description}
}
\value{
list of dataframes containing nodes and edges information
}
\description{
Creates dataframe that describe the network.
The degree and community of each node are computed.
Communities are defined by the louvain algorithm, that 
maximizes local modularity. The gene type is asssigned to
eacg gene, either a regulator or a target gene.
}
\examples{
\dontrun{
data("abiotic_stresses")
data("regulators_per_organism")
genes <- get_locus(abiotic_stresses$heat_DEGs)
# mat was inferred using the function network_inference
mat <- abiotic_stresses$heat_DEGs_regulatory_links
network <- DIANE::network_thresholding(mat, n_edges = length(genes))
data <- network_data(network, regulators_per_organism[["Arabidopsis thaliana"]])
}
}
