% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_edge_testing.R
\name{network_from_tests}
\alias{network_from_tests}
\title{Create network from edges statistical tests}
\usage{
network_from_tests(links, fdr)
}
\arguments{
\item{links}{dataframe of the network edges and associated pvalues,
as in the links attribute of the \code{test_edges} method.}

\item{fdr}{threshold value such as all edges wit hadjusted pvalues lesser than
the argument will be discarded for the final network construction}
}
\value{
an oriented weighted network as an igraph object
}
\description{
Create network from edges statistical tests
}
