% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_inference.R
\name{network_thresholding}
\alias{network_thresholding}
\title{Thresholds a non sparse adjascency matrix to return the strongest weights only}
\usage{
network_thresholding(mat, n_edges)
}
\arguments{
\item{mat}{matrix containing the importance values for each target and regulator}

\item{n_edges}{number of edges to keep in the final network}
}
\value{
igraph object
}
\description{
Thresholds a non sparse adjascency matrix to return the strongest weights only
}
