% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_network_inference.R
\name{network_thresholding}
\alias{network_thresholding}
\title{Thresholds a non sparse adjascency matrix to return the strongest weights only}
\usage{
network_thresholding(mat, n_edges)
}
\arguments{
\item{mat}{matrix containing the importance values for each target and regulator}

\item{n_edges}{number of edges top edges (based on their weight) to keep in the final network}
}
\value{
igraph object
}
\description{
Without thresholding, we would obtain a fully connected weighted 
graph from GENIE3, with far too many links to be interpretable.
In order build a meaningful network, this weighted adjacency matrix between 
regulators and targets has to be sparsified, and we have to determine the regulatory 
weights that we consider significant.
}
\examples{
\dontrun{
data("abiotic_stresses")
data("regulators_per_organism")

# mat was inferred using the function network_inference
mat <- abiotic_stresses$heat_DEGs_regulatory_links
network <- DIANE::network_thresholding(mat, n_edges = length(genes))
}
}
