% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalisation.R
\name{normalize}
\alias{normalize}
\title{normalize raw count data, using the TCC package}
\usage{
normalize(
  data,
  conditions,
  norm_method = "tmm",
  deg_method = "edgeR",
  fdr = 0.01,
  iteration = TRUE
)
}
\arguments{
\item{data}{raw counts to be normalized (data frame or matrix)}

\item{conditions}{condition of each column of the data argument}

\item{norm_method}{method used for normalisation, between tmm or deseq2}

\item{deg_method}{method used for deg detection, between edgeR ou deseq2}

\item{fdr}{pvalue threshold for adjusted pvalues for degs}

\item{iteration}{weather or not to perform a prior removal of DEGs (TRUE or FALSE)}
}
\value{
a TCC-Class object containing the normalized data (filtering is highly recommended after normalization)
}
\description{
normalize raw count data, using the TCC package
}
\examples{
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)
}
