% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{pca_plot_correlation}
\alias{pca_plot_correlation}
\title{PCA plot correlation}
\usage{
pca_plot_correlation(pca, design = NULL, plotRsquared = FALSE)
}
\description{
Draw correlation of each conditions groups to the 
principal components. This is done using the great CorLevelPlot package, by
Kevin Blighe (https://github.com/kevinblighe/CorLevelPlot).
}
\examples{
data("abiotic_stresses")
pca <- compute_pca(abiotic_stresses$normalized_counts)
pca_plot_correlation(pca, abiotic_stresses$design)
}
