% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{pca_plot_correlation}
\alias{pca_plot_correlation}
\title{PCA plot correlation}
\usage{
pca_plot_correlation(pca, design = NULL, plotRsquared = FALSE)
}
\arguments{
\item{pca}{PCA data obtained from compute_pca function.}

\item{design}{experimental design as a dataframe}
}
\description{
Draw correlation of each conditions groups to the 
principal components.
}
\examples{
data("abiotic_stresses")
pca <- compute_pca(abiotic_stresses$normalized_counts)
pca_plot_correlation(pca, abiotic_stresses$design)
}
