% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_viz.R
\name{quick_pca}
\alias{quick_pca}
\title{Quick PCA}
\usage{
quick_pca(data)
}
\arguments{
\item{data}{normalized expression data with samples as columns and genes as rows.}
}
\description{
Draws variables contributions to principal components,
as well as the PCA screeplot.
First to fourth principal components are shown.
}
\examples{
data("abiotic_stresses")
quick_pca(abiotic_stresses$normalized_counts)
}
