% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  server_version = FALSE,
  seed = round(runif(n = 1, min = 0, max = 2^7)),
  host = "127.0.0.1",
  ...
)
}
\arguments{
\item{server_version}{TRUE if the app is deployed on web server,
else (default), FALSE}

\item{seed}{seed for random state to ensure reproducible runs
along DIANE's pipeline}

\item{host}{IP address of the app 
(e.g "127.0.0.1" for localhost, "0.0.0.0", ...)}

\item{...}{A series of options to be used inside the app.}
}
\value{
shiny application
}
\description{
Run the Shiny Application
}
