% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{run_coseq}
\alias{run_coseq}
\title{run_coseq}
\usage{
run_coseq(conds, genes, data, K = 6:12)
}
\arguments{
\item{conds}{Condition names to be used for clustering}

\item{genes}{Genes used as an input for the clustering}

\item{data}{normalized counts (MUST be normalized!)}

\item{K}{number of clusters range}
}
\value{
named list containing the coseq run result as "model", and the cluster membership 
for each gene as "membership"
}
\description{
run_coseq
}
\examples{
\dontrun{
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)
threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)
topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes
clustering <- DIANE::run_coseq(conds = unique(demo_data_At$conditions), 
data = normalized_counts, genes = genes, K = 6:9)
}
}
