% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_coseq.R
\name{run_coseq}
\alias{run_coseq}
\title{run_coseq}
\usage{
run_coseq(conds, genes, data, K = 6:12)
}
\arguments{
\item{conds}{Condition names to be used for clustering}

\item{genes}{Genes used as an input for the clustering}

\item{data}{normalized counts}

\item{K}{number of clusters range}
}
\value{
named list containing the coseq run result, and the cluster membership for each gene
}
\description{
run_coseq
}
